requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-frontmatter>=1.1.0
